refactoring mdlwr
